Extract WordChoice component from Pregame.
